Add files via upload
I wanted to see who were the top importers and exporters of coffee globally, delving into the evolution of their trade patterns from 1990 to 2019. Coffee has been a major influence on my life and for so many around the world. It is fascinating to learn about its influence on people, cultures, and economic structures across the world and I enjoyed exploring this data spanning over the past three decades.