Add files via upload
I will be working with a fake advertising data set, indicating whether or not a particular internet user clicked on an Advertisement on a company website. I will create a model that will predict whether or not they will click on an ad based off the features of that user.